+++ workspace/marqua/package.json
@@ -23,6 +23,7 @@
     ".": "./core/index.js",
     "./artisan": "./artisan/index.js",
     "./browser": "./browser/index.js",
+    "./fs": "./fs/index.js",
     "./styles/code.css": "./styles/code.css",
     "./package.json": "./package.json"
   },
